feat(EditComponent): add max_file_size and content_types to upload buttons inspired by dsfr
