+++ app/assets/stylesheets/utils.scss
@@ -29,15 +29,19 @@
 }
 
 .text-sm {
-  font-size: 14px;
+  font-size: 14px !important;
 }
 
 .text-lg {
   font-size: 18px;
 }
 
-.bold {
-  font-weight: bold;
+.bold-weight-bold {
+  font-weight: bold !important;
+}
+
+.font-weight-normal {
+  font-weight: normal !important;
 }
 
 .numbers-delimiter {
