+++ app/components/attachment/edit_component/edit_component.en.yml
@@ -1,2 +1,3 @@
 ---
 en:
+  max_file_size: "File size limit : %{max_file_size}."
