+++ app/components/attachment/edit_component/edit_component.fr.yml
@@ -1,2 +1,3 @@
 ---
 fr:
+  max_file_size: "Taille maximale : %{max_file_size}."
