+++ app/components/attachment/edit_component/edit_component.html.haml
@@ -24,4 +24,8 @@
       %span.icon.retry
       Ré-essayer
 
+
+  %label.text-sm.font-weight-normal{ for: file_field_options[:id] }
+    = t('.max_file_size', max_file_size: number_to_human_size(max_file_size))
+
   = form.file_field(file_field_name, **file_field_options)
