+++ app/views/administrateurs/procedures/_informations.html.haml
@@ -20,7 +20,7 @@
   = f.select :zone_id, grouped_options_for_zone
 
 %h3.header-subsection Logo de la démarche
-= render Attachment::EditComponent.image(f, @procedure.logo)
+= render Attachment::EditComponent.new(form: f, attached_file: @procedure.logo, direct_upload: true, user_can_destroy: true)
 
 %h3.header-subsection Conservation des données
 = f.label :duree_conservation_dossiers_dans_ds do
@@ -55,7 +55,7 @@
 = f.text_field :cadre_juridique, class: 'form-control', placeholder: 'https://www.legifrance.gouv.fr/'
 
 = f.label :deliberation, 'Importer le texte'
-= render Attachment::EditComponent.text(f, @procedure.deliberation)
+= render Attachment::EditComponent.new(form: f, attached_file: @procedure.deliberation, user_can_destroy: true)
 
 %h3.header-subsection
   RGPD
@@ -84,8 +84,7 @@
 = f.label :notice, 'Notice'
 %p.notice
   Formats acceptés : .doc, .odt, .pdf, .ppt, .pptx
-- notice = @procedure.notice
-= render Attachment::EditComponent.text(f, @procedure.notice)
+= render Attachment::EditComponent.new(form: f, attached_file: @procedure.notice, user_can_destroy: true)
 
 - if !@procedure.locked?
   %h3.header-subsection À qui s’adresse ma démarche ?
