+++ app/views/administrateurs/procedures/publication.html.haml
@@ -44,7 +44,7 @@
               Profitez de la phase de test pour tester la saisie de dossiers, ainsi que toutes les fonctionnalités associées (instruction, emails automatiques, attestations, etc.).
             %p
               Vous pouvez effectuer toutes les modifications que vous souhaitez sur votre démarche pendant cette phase de test.
-            %p.mb-4.bold
+            %p.mb-4.font-weight-bold
               Les dossiers qui seront remplis pendant la phase de test seront automatiquement supprimés lors de la modification ou la publication de votre démarche.
             %p.center
               %iframe{ :src =>"https://player.vimeo.com/video/334463514?color=0069CC",:width =>"640",:height =>"360",:frameborder => "0" }
