+++ app/views/experts/avis/shared/avis/_form.html.haml
@@ -7,7 +7,7 @@
     = f.text_area :introduction, rows: 3, value: avis.introduction || 'Bonjour, merci de me donner votre avis sur ce dossier.', required: true
     %p.tab-title Ajouter une pièce jointe
     .form-group
-      = render Attachment::EditComponent.text(f, avis.introduction_file)
+      = render Attachment::EditComponent.new(form: f, attached_file: avis.introduction_file, user_can_destroy: true)
 
     - if linked_dossiers.present?
       = f.check_box :invite_linked_dossiers, {}, true, false
