+++ app/views/instructeurs/avis/instruction.html.haml
@@ -17,7 +17,7 @@
 
     = form_for @avis, url: instructeur_avis_path(@avis.procedure, @avis), html: { class: 'form' } do |f|
       = f.text_area :answer, rows: 3, placeholder: 'Votre avis', required: true
-      = render Attachment::EditComponent.text(f, @avis.piece_justificative_file)
+      = render Attachment::EditComponent.new(form: f, attached_file: @avis.piece_justificative_file, user_can_destroy: true)
 
       .flex.justify-between.align-baseline
         %p.confidentiel.flex
