+++ app/views/shared/dossiers/messages/_form.html.haml
@@ -8,10 +8,7 @@
     - disable_piece_jointe = defined?(disable_piece_jointe) ? disable_piece_jointe : false
     %div
       - if !disable_piece_jointe
-        = f.label :piece_jointe, for: :piece_jointe do
-          = t('views.shared.dossiers.messages.form.attach_dossier')
-          %span.notice= t('views.shared.dossiers.messages.form.attachment_size')
-        = f.file_field :piece_jointe, id: 'piece_jointe', direct_upload: true
+        = render Attachment::EditComponent.new(form: f, attached_file: commentaire.piece_jointe)
 
     %div
       = f.submit t('views.shared.dossiers.messages.form.send_message'), class: 'button primary send', data: { disable: true }
